Bluetooth service working + Splash screen
