+++ .idea/misc.xml
@@ -29,6 +29,7 @@
         <entry key="..\:/Users/Julian Chan/OneDrive/Documents/SCHOOL/2022_SUMMER/CMPT_362/Project/app/src/main/res/xml/preference.xml" value="0.5369791666666667" />
         <entry key="..\:/Users/Julian Chan/OneDrive/Documents/SCHOOL/2022_SUMMER/CMPT_362/Project/app/src/main/res/xml/preference_settings.xml" value="0.3328125" />
         <entry key="..\:/Users/Shashank/Desktop/Summer 2022/CMPT362/Project/BlueRadar/app/src/main/res/layout/activity_main.xml" value="0.264" />
+        <entry key="..\:/Users/Shashank/Desktop/Summer 2022/CMPT362/Project/BlueRadar/app/src/main/res/layout/activity_splash_screen.xml" value="0.4036458333333333" />
         <entry key="..\:/Users/Shashank/Desktop/Summer 2022/CMPT362/Project/BlueRadar/app/src/main/res/layout/app_bar_main.xml" value="0.40260416666666665" />
         <entry key="..\:/Users/Shashank/Desktop/Summer 2022/CMPT362/Project/BlueRadar/app/src/main/res/layout/burger_nav_header.xml" value="0.40260416666666665" />
         <entry key="..\:/Users/Shashank/Desktop/Summer 2022/CMPT362/Project/BlueRadar/app/src/main/res/layout/custom_toolbar.xml" value="0.40260416666666665" />
